rename classes to match with illustration
