+++ hexagonal-architecture/src/main/java/com/baeldung/hexagonal/adapters/StockAdapter.java
@@ -1,0 +1,24 @@
+package com.baeldung.hexagonal.adapters;
+
+import com.baeldung.hexagonal.ports.StockPort;
+import com.baeldung.hexagonal.repository.StockRepository;
+
+/**
+ * User: raychenon
+ * Date: 12/7/2020
+ * <p>
+
+ */
+public class StockAdapter implements StockPort {
+
+    private StockRepository stockRepository;
+
+    public StockAdapter(StockRepository stockRepository) {
+        this.stockRepository = stockRepository;
+    }
+    
+    @Override
+    public int[] getStockPrices(String stockName) {
+        return stockRepository.getStock(stockName);
+    }
+}
