+++ hexagonal-architecture/src/main/java/com/baeldung/hexagonal/core/StockPriceCore.java
@@ -1,25 +1,25 @@
 package com.baeldung.hexagonal.core;
 
-import com.baeldung.hexagonal.adapters.BestTimetoBuyandSellStockAdapter;
 import com.baeldung.hexagonal.repository.StockFetcher;
 import com.baeldung.hexagonal.repository.StockRepository;
 
 /**
  * User: raychenon
  * Date: 14/7/2020
+ * No framework dependencies
  */
 public class StockPriceCore {
 
     private StockRepository stockRepository;
-    private BestTimetoBuyandSellStockAdapter adapter;
+    private ProfitCalculator profitCalculator;
 
     public StockPriceCore() {
         this.stockRepository = new StockRepository(new StockFetcher());
-        this.adapter = new BestTimetoBuyandSellStockAdapter();
+        profitCalculator = new ProfitCalculator();
     }
 
     public int getBestPossibleProfit(String stockName) {
-        return adapter.maxProfit(stockRepository.getStock(stockName));
+        return profitCalculator.maxProfit(stockRepository.getStock(stockName));
     }
 
     public int[] getStockPrices(String stockName) {
