+++ hexagonal-architecture/src/main/java/com/baeldung/hexagonal/core/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Everything in core package contains the business logic
+ */
+package com.baeldung.hexagonal.core;
