+++ hexagonal-architecture/src/main/java/com/baeldung/hexagonal/ports/StockPort.java
@@ -1,0 +1,11 @@
+package com.baeldung.hexagonal.ports;
+
+/**
+ * User: raychenon
+ * Date: 12/7/2020
+ */
+public interface StockPort {
+
+    public int[] getStockPrices(String stockName);
+
+}
